src/main/java/com/example/testes/domain/PlanetService.java
package com.example.testes.domain;

import java.util.List;
import java.util.Optional;

import org.springframework.data.domain.Example;
import org.springframework.stereotype.Service;

@Service
public class PlanetService {

    private PlanetRepository planetRepository;

    public PlanetService(PlanetRepository planetRepository) {
        this.planetRepository = planetRepository;
    }

    public Planet create(Planet planet) {
        return planetRepository.save(planet);
    }

    public Optional<Planet> get(Long id) {
        return planetRepository.findById(id);
    }

    public Optional<Planet> getByName(String name) {
        return planetRepository.findByName(name);
    }

    /*
     * Uso da API QueryExample. Isso permite criar querys dinâmicas.
     * É criado uma query com base nos atributos da entidade criada,
     * ao passar esta no método makeQuery.
     * Além disso, é necessário sobreescrever o método findAll no repositório,
     * passando o objeto Example.
     */
    public List<Planet> list(String climate, String terrain) {
        Example<Planet> query = QueryBuilder.makeQuery(new Planet(climate, terrain));
        return planetRepository.findAll(query);
    }

    public void remove(Long id) {
        planetRepository.deleteById(id);
    }

}
